netbsd: wire up the penumbra MD inline byteswaps
penumbra/byte_swap.h has carried correct inline
__BYTE_SWAP_U16/U32_VARIABLE definitions since port bring-up, but
machine/bswap.h never included it, so sys/bswap.h fell back to the
out-of-line libc bswap functions: SHA-256 paid a call per message
word, and htonl/ntohl across kernel and userland compiled to calls.

Include it as <machine/byte_swap.h> (the i386/vax single-arch form —
stand builds alias only machine/, so the arch-name form breaks
stand/boot), replace sys/types.h with sys/cdefs.h + sys/stdint.h to
break an include cycle through sys/endian.h, and cast the u16 swap
result back to uint16_t so integer promotion survives -Werror.

+++ netbsd/sys/arch/penumbra/include/byte_swap.h
@@ -4,14 +4,15 @@
 #define _PENUMBRA_BYTE_SWAP_H_
 
 #ifdef __GNUC__
-#include <sys/types.h>
+#include <sys/cdefs.h>
+#include <sys/stdint.h>
 __BEGIN_DECLS
 
 #define	__BYTE_SWAP_U16_VARIABLE __byte_swap_u16_variable
 static __inline uint16_t
 __byte_swap_u16_variable(uint16_t v)
 {
-	return (v >> 8) | (v << 8);
+	return (uint16_t)((v >> 8) | (v << 8));
 }
 
 #define	__BYTE_SWAP_U32_VARIABLE __byte_swap_u32_variable
